+++ src/ome_zarr_models/v04/__init__.py
@@ -1,5 +1,4 @@
 from ome_zarr_models.v04.hcs import HCS
 from ome_zarr_models.v04.image import Image
-from ome_zarr_models.v04.image2 import Image2
 
-__all__ = ["HCS", "Image", "Image2"]
+__all__ = ["HCS", "Image"]
