+++ tests/v04/conftest.py
@@ -11,7 +11,7 @@ from zarr.util import guess_chunks
 
 from ome_zarr_models.base import Base
 from ome_zarr_models.v04.axes import Axis
-from ome_zarr_models.v04.image import Image, ImageAttrs
+from ome_zarr_models.v04.image_old import Image, ImageAttrs
 from ome_zarr_models.v04.multiscales import (
     Dataset,
     Multiscale,
